Moved upload widget to a new file
